refactor: UserRequest 패키지 이동
