+++ tomcat/src/test/java/org/apache/coyote/http11/QueryParamTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import nextstep.jwp.response.UserRequest;
+import nextstep.jwp.request.UserRequest;
 import org.apache.coyote.http11.exception.QueryParamNotFoundException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
